Initial Commit
This code will make the LEDs light up in sequence from one end to the other and then back again, creating the iconic Knight Rider effect. You can adjust the delay time to change the speed of the effect. Just replace the 100 in the delay() function with a different value to make it faster or slower.